refact: move styles to core.feed.profile.toot.style.type
