+++ core/mastodon/src/main/java/com/jeanbarrossilva/orca/core/mastodon/feed/profile/toot/Status.kt
@@ -3,7 +3,7 @@ package com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot
 import android.util.Log
 import com.jeanbarrossilva.orca.core.feed.profile.toot.style.toStyledString
 import com.jeanbarrossilva.orca.core.mastodon.feed.profile.account.MastodonAccount
-import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.style.styling.mention.MastodonMentionDelimiter
+import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.style.type.MastodonMentionDelimiter
 import java.net.URL
 import java.time.ZonedDateTime
 import kotlinx.serialization.Serializable
