+++ core/mastodon/src/main/java/com/jeanbarrossilva/orca/core/mastodon/feed/profile/toot/style/Element.extensions.kt
@@ -1,6 +1,6 @@
 package com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.style
 
-import com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.mention.Mention
+import com.jeanbarrossilva.orca.core.feed.profile.toot.style.type.Mention
 import org.jsoup.nodes.Element
 
 /** Whether this [Element] is that of a [Mention]. **/
