+++ core/mastodon/src/test/java/com/jeanbarrossilva/orca/core/mastodon/feed/profile/toot/MastodonMentionDelimiterTests.kt
@@ -1,7 +1,7 @@
 package com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot
 
 import com.jeanbarrossilva.orca.core.feed.profile.Profile
-import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.style.styling.mention.MastodonMentionDelimiter
+import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.style.type.MastodonMentionDelimiter
 import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.test.sample
 import com.jeanbarrossilva.orca.core.sample.feed.profile.sample
 import org.junit.Assert.assertEquals
