+++ core/src/main/java/com/jeanbarrossilva/orca/core/feed/profile/toot/style/Style.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling
+package com.jeanbarrossilva.orca.core.feed.profile.toot.style
 
 import java.io.Serializable
