+++ core/src/main/java/com/jeanbarrossilva/orca/core/feed/profile/toot/style/type/Bold.kt
@@ -1,6 +1,6 @@
-package com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.bold
+package com.jeanbarrossilva.orca.core.feed.profile.toot.style.type
 
-import com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.Style
+import com.jeanbarrossilva.orca.core.feed.profile.toot.style.Style
 
 /** [Style] that applies a heavier weight to the target's font. **/
 data class Bold(override val indices: IntRange) : Style() {
